refactor: Move all server logic to bot.js, remove unnecessary code.
Move ability to send messages to Discord to bot.js so that it is the only file that deals with the
Discord.js library. Remove any unused variables. Change the way the response is created so that it
now shows the discarded rolls that are the result of rolling with advantage or disadvantage.